Centralize running wasi test
