+++ crates/containerd-shim-wasm/Cargo.toml
@@ -28,12 +28,12 @@ clone3 = "0.2"
 libc = { workspace = true }
 caps = "0.5"
 proc-mounts = "0.3"
+tempfile = "3"
 
 [build-dependencies]
 ttrpc-codegen = { version = "0.4.2", optional = true }
 
 [dev-dependencies]
-tempfile = "3"
 pretty_assertions = "1"
 signal-hook = "0.3"
 env_logger = { workspace = true }
