+++ crates/containerd-shim-wasm/src/sandbox/testutil.rs
@@ -1,7 +1,22 @@
 //! Testing utilities used across different modules
 
-use super::{Error, Result};
-use std::process::{Command, Stdio};
+use super::{instance::Wait, EngineGetter, Error, Instance, InstanceConfig, Result};
+use anyhow::Result as AnyHowResult;
+use chrono::{DateTime, Utc};
+use libc::SIGKILL;
+use oci_spec::runtime::{ProcessBuilder, RootBuilder, SpecBuilder};
+use serde::{Deserialize, Serialize};
+use std::io::Write;
+use std::os::unix::fs::OpenOptionsExt;
+use std::{
+    borrow::Cow,
+    fs::{create_dir, File, OpenOptions},
+    path::PathBuf,
+    process::{Command, Stdio},
+    sync::mpsc::channel,
+    time::Duration,
+};
+use tempfile::TempDir;
 
 fn normalize_test_name(test: &str) -> Result<&str> {
     let closure_removed = test.trim_end_matches("::{{closure}}");
@@ -64,5 +79,91 @@ macro_rules! function {
         name
     }};
 }
-
 pub use function;
+
+#[derive(Serialize, Deserialize)]
+struct Options {
+    root: Option<PathBuf>,
+}
+
+pub fn run_wasi_test<I, E>(
+    dir: &TempDir,
+    wasmbytes: Cow<[u8]>,
+    start_fn: Option<&str>,
+) -> AnyHowResult<(u32, DateTime<Utc>), Error>
+where
+    I: Instance<E = E> + EngineGetter<E = E>,
+    E: Sync + Send + Clone,
+{
+    create_dir(dir.path().join("rootfs"))?;
+    let rootdir = dir.path().join("runwasi");
+    create_dir(&rootdir)?;
+    let opts = Options {
+        root: Some(rootdir),
+    };
+    let opts_file = OpenOptions::new()
+        .read(true)
+        .create(true)
+        .truncate(true)
+        .write(true)
+        .open(dir.path().join("options.json"))?;
+    write!(&opts_file, "{}", serde_json::to_string(&opts)?)?;
+
+    let wasm_path = dir.path().join("rootfs/hello.wasm");
+    let mut f = OpenOptions::new()
+        .write(true)
+        .create(true)
+        .truncate(true)
+        .mode(0o755)
+        .open(wasm_path)?;
+    f.write_all(&wasmbytes)?;
+
+    let stdout = File::create(dir.path().join("stdout"))?;
+    drop(stdout);
+
+    let entrypoint = match start_fn {
+        Some(s) => "./hello.wasm#".to_string() + s,
+        None => "./hello.wasm".to_string(),
+    };
+    let spec = SpecBuilder::default()
+        .root(RootBuilder::default().path("rootfs").build()?)
+        .process(
+            ProcessBuilder::default()
+                .cwd("/")
+                .args(vec![entrypoint])
+                .build()?,
+        )
+        .build()?;
+
+    spec.save(dir.path().join("config.json"))?;
+
+    let mut cfg = InstanceConfig::new(
+        I::new_engine()?,
+        "test_namespace".into(),
+        "/containerd/address".into(),
+    );
+    let cfg = cfg
+        .set_bundle(dir.path().to_str().unwrap().to_string())
+        .set_stdout(dir.path().join("stdout").to_str().unwrap().to_string());
+
+    let wasi = I::new("test".to_string(), Some(cfg));
+
+    wasi.start()?;
+
+    let (tx, rx) = channel();
+    let waiter = Wait::new(tx);
+    wasi.wait(&waiter).unwrap();
+
+    let res = match rx.recv_timeout(Duration::from_secs(10)) {
+        Ok(res) => Ok(res),
+        Err(e) => {
+            wasi.kill(SIGKILL as u32).unwrap();
+            return Err(Error::Others(format!(
+                "error waiting for module to finish: {0}",
+                e
+            )));
+        }
+    };
+    wasi.delete()?;
+    res
+}
