+++ crates/containerd-shim-wasmtime/Cargo.toml
@@ -42,6 +42,7 @@ tempfile = "3.7"
 libc = { workspace = true }
 pretty_assertions = "1"
 env_logger = "0.10"
+serial_test = "*"
 
 [[bin]]
 name = "containerd-shim-wasmtime-v1"
